% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSWO-package.R
\docType{package}
\name{NPSWO-package}
\alias{NPSWO}
\alias{NPSWO-package}
\title{NPSWO: MSE Framework for North Pacific Swordfish}
\description{
A framework for conducting Management Strategy Evaluation (MSE) for North Pacific swordfish (Xiphias gladius). Built on the openMSE ecosystem, with MSEtool serving as the primary MSE engine. The package provides functions specific to the North Pacific swordfish MSE, including operating model conditioning, stochastic life-history parameter sampling, and Stock Synthesis 3 (SS3) model setup and execution. The root directory of the GitHub repository contains all code required to reproduce the MSE.
}
\author{
\strong{Maintainer}: Adrian Hordyk \email{adrian@bluematterscience.com}

Authors:
\itemize{
  \item Rebecca Scott \email{rscott@bluematterscience.com}
}

}
\keyword{internal}
